// apps/cofepris/src/app/application/tramites/260910/models/asociados.model.ts
/**
 * Interfaz que representa la información de un fabricante.
 */
export interface Asociados {
  /** Nombre del Folio Tramite. */
  folioTramite: string;

  /** Nombre del Tipo Tramite. */
  tipoTramite: string;

  /** Nombre del Estatus. */
  estatus: string;

  /** Nombre del Fecha Alta De Registro. */
  fechaAltaDeRegistro: string;
  
}